+++ packages/components/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Internal
 
 -   Refactor `FocalPointPicker` to function component ([#39168](https://github.com/WordPress/gutenberg/pull/39168)).
+-   `Guide`: use `code` instead of `keyCode` for keyboard events ([#43604](https://github.com/WordPress/gutenberg/pull/43604/)).
 
 ## 20.0.0 (2022-08-24)